BAEL-1980 Access an in-memory h2 db from 2 Spring Boot applications (#5230)
* Added Spring-Boot-H2 project

* Optimized imports and formatted code

* Code Formatting changes

* Renamed the client applicaiton

* Removed Name from pom.xml

* formatting changes - applied formatting style

* added souts

* Update SpringBootApp.java

* Update ClientSpringBootApp.java

* Fixed Formatting issues

* Fixed formatting and spelling mistakes

* added H2 shared db example

* a

* patch

* test

* Revert "Fixed formatting and spelling mistakes"

This reverts commit ed56ededec903e85c2d9f2b55e7730682455d5f5.

* Revert "test"

This reverts commit 4a7133dbd1d181cc304348f63998924a9db5acfe.

* Revert "patch"

This reverts commit 404b9ebe74aef4faa2ad62d9867652b62a5445b8.

* Revert "a"

This reverts commit af77f841114e19b0a0eafb15530dbaaaf34b5493.

* Revert "added H2 shared db example"

This reverts commit e16918cae3813fe20172c7e90e27bbec34b5fb5f.

* Updated to have client and server in same directory

* Deleted old files

* removed unwanted directory

* formating

* removed unwanted module

* commentted removed module

* Revert "commentted removed module"

This reverts commit 211ff1441f6bde920d7af3f8cd52c2a682ba3967.

* Revert "removed unwanted module"

This reverts commit 657f9ee45e02269bcfdbf380b9a74ff2370857fc.

* fixed pom.xml issue

* typo fixed

* fixed multiple main class issue
