+++ spring-boot-h2/spring-boot-h2-database/pom.xml
@@ -22,6 +22,8 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <java.version>1.8</java.version>
+        <!-- The main class to start by executing java -jar -->
+        <start-class>com.mycorp.starter.HelloWorldApplication</start-class>
     </properties>
 
     <dependencies>
