+++ spring-boot-h2/spring-boot-h2-database/src/main/java/com/baeldung/h2db/demo/server/SpringBootApp.java
@@ -1,31 +1,38 @@
-package com.baeldung.h2db.demo;
+package com.baeldung.h2db.demo.server;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.Arrays;
 import javax.annotation.PostConstruct;
+import org.h2.tools.Server;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 
 @SpringBootApplication
-public class ClientSpringBootApp {
+@ComponentScan("com.baeldung.h2db.demo.server")
+public class SpringBootApp {
 
     @Autowired
     private JdbcTemplate jdbcTemplate;
 
     public static void main(String[] args) {
-        SpringApplication.run(ClientSpringBootApp.class, args);
+        SpringApplication.run(SpringBootApp.class, args);
     }
 
     @PostConstruct
     private void initDb() {
-        System.out.println("****** Inserting more sample data in the table: Employees ******");
+        System.out.println(String.format("****** Creating table: %s, and Inserting test data ******", "Employees"));
+
         String sqlStatements[] = {
-            "insert into employees(first_name, last_name) values('Donald','Trump')",
-            "insert into employees(first_name, last_name) values('Barack','Obama')"
+            "drop table employees if exists",
+            "create table employees(id serial,first_name varchar(255),last_name varchar(255))",
+            "insert into employees(first_name, last_name) values('Eugen','Paraschiv')",
+            "insert into employees(first_name, last_name) values('Scott','Tiger')"
         };
 
         Arrays.asList(sqlStatements).stream().forEach(sql -> {
@@ -45,5 +52,10 @@ public class ClientSpringBootApp {
                     return null;
                 }
             });
+    }
+
+    @Bean(initMethod = "start", destroyMethod = "stop")
+    public Server inMemoryH2DatabaseServer() throws SQLException {
+        return Server.createTcpServer("-tcp", "-tcpAllowOthers", "-tcpPort", "9091");
     }
 }
