--- spring-boot-h2/spring-boot-h2-remote-app/src/main/resources/application.properties
@@ -1,5 +1,0 @@
-spring.datasource.url=jdbc:h2:tcp://localhost:9091/mem:mydb
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=
-spring.jpa.hibernate.ddl-auto=create
